+++ src/components/app/app.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {Switch, Route, BrowserRouter} from 'react-router-dom';
 import MainPage from '../main/main';
-import SignIn from '../login/SignIn';
-import MyList from '../mylist/MyList';
-import Film from '../films/id/Film';
-import AddReview from '../films/id/review/AddReview';
-import Player from '../player/id/Player';
-import NotFoundPage from '../NotFoundPage';
+import SignIn from '../login/login';
+import MyList from '../my-list/my-list';
+import Film from '../film/film';
+import AddReview from '../add-review/add-review';
+import Player from '../player/player';
+import NotFoundPage from '../not-found-page/not-found-page';
 
 const App = (props) => {
 
@@ -29,13 +29,13 @@ const App = (props) => {
         <Route exact path="/mylist">
           <MyList />
         </Route>
-        <Route exact path="/films/id">
+        <Route exact path="/film">
           <Film />
         </Route>
-        <Route exact path="/films/id/review">
+        <Route exact path="/review">
           <AddReview />
         </Route>
-        <Route exact path="/player/id">
+        <Route exact path="/player">
           <Player />
         </Route>
         <Route>
